Правки кода 17.04.23 Разделение проекта по папкам было преобразовано. Зависимости проекта автоматически перезаписались, кода изменено не было вообще
