+++ module.plant-service/src/main/java/ru/botanica/dtos/PlantDto.java
@@ -1,10 +1,9 @@
-package ru.botanica.entities.plants;
+package ru.botanica.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.stereotype.Component;
-import ru.botanica.entities.plantCares.PlantCareDto;
 
 import java.util.List;
