+++ module.plant-service/src/main/java/ru/botanica/entities/Care.java
@@ -1,4 +1,4 @@
-package ru.botanica.entities.care;
+package ru.botanica.entities;
 
 import jakarta.persistence.*;
 import lombok.Getter;
