+++ module.plant-service/src/main/java/ru/botanica/entities/PlantPhoto.java
@@ -1,4 +1,4 @@
-package ru.botanica.entities.photos;
+package ru.botanica.entities;
 
 import jakarta.persistence.*;
 import lombok.Getter;
