+++ module.plant-service/src/main/java/ru/botanica/repositories/CareRepository.java
@@ -3,7 +3,7 @@ package ru.botanica.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
-import ru.botanica.entities.care.Care;
+import ru.botanica.entities.Care;
 
 @Repository
 public interface CareRepository extends JpaRepository<Care, Long>, JpaSpecificationExecutor<Care> {
