+++ module.plant-service/src/main/java/ru/botanica/repositories/PlantPhotoRepository.java
@@ -3,7 +3,7 @@ package ru.botanica.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Repository;
-import ru.botanica.entities.photos.PlantPhoto;
+import ru.botanica.entities.PlantPhoto;
 
 @Repository
 public interface PlantPhotoRepository extends JpaRepository<PlantPhoto, Long> {
