+++ module.plant-service/src/main/java/ru/botanica/repositories/PlantRepository.java
@@ -3,7 +3,7 @@ package ru.botanica.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
-import ru.botanica.entities.plants.Plant;
+import ru.botanica.entities.Plant;
 
 import java.util.Optional;
