+++ module.plant-service/src/test/java/ru/botanica/services/PlantControllerTests.java
@@ -8,8 +8,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import ru.botanica.controllers.PlantController;
-import ru.botanica.entities.photos.PlantPhoto;
-import ru.botanica.entities.plants.Plant;
+import ru.botanica.entities.PlantPhoto;
+import ru.botanica.entities.Plant;
 import ru.botanica.repositories.PlantRepository;
 
 import java.util.List;
